iot-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-generativeai==0.8.3
gTTS==2.5.1
websockets==15.0.1
httpx==0.28.1
python-dotenv==1.0.0
python-dateutil==2.8.2
pydantic-settings==2.10.1
email-validator==2.3.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.28.1
pytest-mock==3.12.0
# install with pip install -r requirements.txt